UML programmed
Not sure if this is correct, but this is the UML represented in the Program.cs file. I am not sure if this is where you wanted to put all of this or not, but at least it will give you a head start. If not, just reject it, I am ok with that.